+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `pynopegl` log levels are now controled using the `pynopegl.Log` enum
 - `max_texture_dimensions_*` capabilities are renamed to `max_texture_dimension_*`
 - Backend probing in `pynopegl` now returns a more Pythonic output
+- `GraphicConfig.scissor` is now an `ivec4` parameter instead of `vec4`
 
 ### Removed
 - `ResourceProps.variadic` bool flag as it was never a functional interface